fix(tests): move RFC-0099 coverage to the roster and close the ledger
CI's pack-test boundary lint reported 24 failures: both new suites read above
`packs/core` — the RFC answer key, `docs/specs/`, and other packs' contracts —
which `tools/lint-pack-test-boundary.py` forbids from inside a pack's test
tree. `pytest` never notices, so this only surfaced in `Caps enforcer
self-test`, `gate-main`, and `make build-check`.

- Split rather than relocate wholesale. `test_routing_precedence.py` returns to
  its upstream state, keeping its four pack-local router contracts anchored
  inside `packs/core`.
- The R1-R12 activation coverage moves to
  `tests/roster/test_rfc0099_activation_coverage.py`, and the fixture-register
  integrity contract moves to `tests/roster/`, both re-anchored at the
  repository root. Collection is unchanged at 23 cases (15 + 4 + 4).
- Wire both by name into `build-check.yml` with a `lint-ci-parity`
  disposition. CI does not auto-discover roster tests, so an unwired suite
  would gate locally and never once remotely. Step count 78 -> 79.
- Repoint the register's twelve ACT evidence paths and the spec's AC2 at the
  new location, and record why the coverage is repository-level.

Also, at the maintainer's request:

- Append the Approver-signed Errata entry recording follow-on #6 discharged and
  six of six discharged. The header stays `Accepted`: the declared vocabulary
  has no shipped status. The entry discharges the delivery ledger, not D7 — the
  adopter study was waived and never run.
- Correct the stale `rfc-architecture-simplification` row in the specs index,
  which read `Approved` while that spec and `workspace.toml` both said
  `Shipped`.

Verified: boundary lint 24 failures -> 0; lint-ci-parity 84 steps all
dispositioned; 23 tests pass; ruff, mypy, and `SKIP_SAST=1 make build-check`
exit 0 after clearing a stale `dist/`. The SAST leg is left to CI.
