+++ docs/rfc/0099-cut-before-adding-and-artifact-shaping.md
@@ -1077,3 +1077,19 @@ record. Corrections are appended here, Approver-signed.
   instrument for any later owner who chooses to run a new study. Follow-on #6
   closes when the two missing executable fixture families and the five-field
   fixture register are complete; no adopter-study artifact is owed.
+- **2026-09-01 (Approver: eugenelim) — follow-on #6 is discharged; six of six
+  are now discharged and the status field stays `Accepted`.** The two closing
+  conditions set by the 2026-08-31 waiver entry are met. The missing executable
+  fixture families are complete: R1–R12 activation coverage across all eight
+  entry points is bound cell by cell to the normative routing table, and the six
+  seeded shaping-review efficacy defects carry independently adjudicated,
+  digest-bound receipts. The versioned five-field fixture register that section
+  6 required before acceptance now exists and covers all seven families —
+  `ACT`, `SHAPE`, `ALIAS`, `RFCARCH`, `STATE`, `CORE`, and `BOUNDARY` — citing
+  shipped evidence in place for the five families already discharged per-spec.
+  This RFC is therefore delivered and auditable end to end. The declared
+  vocabulary still has no implemented or shipped status, so `Accepted` remains
+  terminal and the header does not move. What is discharged is the delivery
+  ledger, not D7: the adopter routing study was waived and never run, its
+  question stays unanswered, and no entry here claims the study passed or that
+  the activation fixtures answer it.
