+++ docs/specs/README.md
@@ -30,7 +30,7 @@ docs/specs/<feature>/
 | [`direct-skill-repository-installation/`](direct-skill-repository-installation/spec.md) | Approved | RFC-0098; RFC-0085; RFC-0031; RFC-0034; ADR-0036; ADR-0039; ADR-0100 | **Admits bounded direct skill repositories through the canonical AgentBundle install lifecycle.** Exact candidate-rooted shape classification, SHA-pinned GitHub acquisition, transport and measured-envelope bounds, fail-closed admission, manifest and lazy state compatibility, lifecycle, CLI, security evidence, and author documentation preserve existing catalogue behavior while refusing unsupported transports and broad discovery. Full mode; 36 ACs / 20 tasks, opening with four spikes. |
 | [`shaping-review-contracts/`](shaping-review-contracts/spec.md) | Shipped | RFC-0099; ADR-0042 | One cold, stateless, read-only shaping reviewer for intent, delivery-brief, and spec contracts; skills and repository content remain the knowledge surfaces, Core MCP is only an invocation route, and `new-spec` keeps acceptance claims atomic and plans grounded rather than guessed. |
 | [`rfc0099-migration-validation-record/`](rfc0099-migration-validation-record/spec.md) | Shipped | RFC-0099 | Waives the unrun adopter routing study without claiming usability validation, closes activation and shaping-review efficacy fixture gaps, and publishes the maintained five-field register for all seven fixture families. |
-| [`rfc-architecture-simplification/`](rfc-architecture-simplification/spec.md) | Approved | RFC-0099 | Pre-write RFC artifact choice, adversarial RFC YAGNI review, architecture reuse/Stage-0 stopping, and architecture-specific YAGNI review without a new reviewer. |
+| [`rfc-architecture-simplification/`](rfc-architecture-simplification/spec.md) | Shipped | RFC-0099 | Pre-write RFC artifact choice, adversarial RFC YAGNI review, architecture reuse/Stage-0 stopping, and architecture-specific YAGNI review without a new reviewer. |
 | [`local-pytest-process-optimization/`](local-pytest-process-optimization/spec.md) | Shipped | ADR-0071; ADR-0094; ADR-0096; `local-ci-shared-test-deduplication` | **Folds four collection floors into their real pytest executions and consolidates only one proven five-file root/tool compatibility class.** Approved root/tool topology is 19→15 standalone and 18→14 composed; pack boundaries, Opportunity-1 ownership, all ambiguous singletons, and the specialized import-path child remain unchanged. Full mode; 22 ACs / 3 tasks complete. |
 | [`okf-follow-ons/`](okf-follow-ons/spec.md) | Implementing | RFC-0087; ADR-0093; PR #1130 follow-ons | **Closes three deliberately deferred OKF defects:** bounds and escapes compiler-owned index display metadata, mutation-pins `OKF012`, and restores architect discovery/source-root accuracy with independent pack releases. Full mode; mixed shape; 9 ACs / 5 tasks. |
 | [`local-ci-shared-test-deduplication/`](local-ci-shared-test-deduplication/spec.md) | Shipped | ADR-0017; ADR-0086; ADR-0096; `local-ci-orchestration` | **Runs the five proven shared test files once inside one composed `make ci` while leaving standalone build-check, test, and make-free Windows routes complete.** Build-check keeps ownership and ordering; an explicit dependency-safe reduced test target carries no ambient profile or persistent state. Accepted ADR-0096 licenses status-only partial-supersession pointers on the frozen local-CI artifacts. Full mode; 18 ACs / 3 tasks complete. |
