+++ docs/specs/rfc0099-migration-validation-record/fixture-register.md
@@ -11,18 +11,18 @@ Existing shipped fixture families are cited in place; they are not rebuilt.
 
 | Fixture ID | Prompt or seeded defect | Installed profile | Exact expected result | Owner |
 | --- | --- | --- | --- | --- |
-| ACT-R1 | “Capture this stated product outcome as a repository intent.” | Core-only | `intake-intent`; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R2 | “Shape this raw product idea before deciding its repository artifact.” | Core + Product Engineering | `frame-intent`, then `intake-intent` only when repository admission is requested; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R3 | “Start this Jira Story”; one shippable behavior and no named artifact. | Core-only | `work-intake` delegates to `new-spec` after classification as one public route; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R4 | “Create a spec for this already-clear behavior.” | Core-only | `new-spec` directly; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R5 | “Draft an RFC for this unresolved consequential direction.” | Governance | `new-rfc` directly with no intent; profile-gated evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Governance pack maintainers |
-| ACT-R6 | “How should we design this integration? Two viable technical shapes remain.” | Architecture | `architect-design` directly; profile-gated evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Architect pack maintainers |
-| ACT-R7 | “Turn this external multi-artifact brief into a repository brief.” | Core-only | `author-delivery-brief create`; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R8 | “Continue this Ready brief, but select no delivery slice yet.” | Core-only | `author-delivery-brief continue`, then stop without `new-spec`; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R9 | “Propose the minimum spec cut from this Ready brief.” | Core-only | `author-delivery-brief continue`; invoke `new-spec` only after separate confirmation; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R10 | “The established behavior regressed; diagnose and fix it.” | Core-only | `bug-fix` directly; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R11 | “Refresh this accepted tracker-origin artifact.” | Core-only | `work-intake` refresh boundary with the resolved processor internal; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
-| ACT-R12 | “What is ready to work on?” | Core-only | `workspace-status` directly; evidence: `packs/core/tests/skills/work-intake/test_routing_precedence.py` | Core work-intake maintainers |
+| ACT-R1 | “Capture this stated product outcome as a repository intent.” | Core-only | `intake-intent`; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R2 | “Shape this raw product idea before deciding its repository artifact.” | Core + Product Engineering | `frame-intent`, then `intake-intent` only when repository admission is requested; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R3 | “Start this Jira Story”; one shippable behavior and no named artifact. | Core-only | `work-intake` delegates to `new-spec` after classification as one public route; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R4 | “Create a spec for this already-clear behavior.” | Core-only | `new-spec` directly; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R5 | “Draft an RFC for this unresolved consequential direction.” | Governance | `new-rfc` directly with no intent; profile-gated evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Governance pack maintainers |
+| ACT-R6 | “How should we design this integration? Two viable technical shapes remain.” | Architecture | `architect-design` directly; profile-gated evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Architect pack maintainers |
+| ACT-R7 | “Turn this external multi-artifact brief into a repository brief.” | Core-only | `author-delivery-brief create`; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R8 | “Continue this Ready brief, but select no delivery slice yet.” | Core-only | `author-delivery-brief continue`, then stop without `new-spec`; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R9 | “Propose the minimum spec cut from this Ready brief.” | Core-only | `author-delivery-brief continue`; invoke `new-spec` only after separate confirmation; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R10 | “The established behavior regressed; diagnose and fix it.” | Core-only | `bug-fix` directly; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R11 | “Refresh this accepted tracker-origin artifact.” | Core-only | `work-intake` refresh boundary with the resolved processor internal; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
+| ACT-R12 | “What is ready to work on?” | Core-only | `workspace-status` directly; evidence: `tests/roster/test_rfc0099_activation_coverage.py` | Core work-intake maintainers |
 | SHAPE-UNNECESSARY-INTENT | A new intent and wrapper skill duplicate `workspace-status` in `docs/specs/rfc0099-migration-validation-record/fixtures/intent-fast-status.md`. | Core-only | Fresh intent-mode review identifies the unnecessary intent as `SHAPE-UNNECESSARY-INTENT`; adjudicated evidence: `docs/specs/rfc0099-migration-validation-record/notes/shaping-efficacy-qa.md` | Core shaping-reviewer maintainers |
 | SHAPE-UNSAFE-SIMPLIFICATION | Confinement and reconciliation are deferred in `docs/specs/rfc0099-migration-validation-record/fixtures/intent-fast-status.md`. | Core-only | Fresh intent-mode review identifies unsafe simplification as `SHAPE-UNSAFE-SIMPLIFICATION`; adjudicated evidence: `docs/specs/rfc0099-migration-validation-record/notes/shaping-efficacy-qa.md` | Core shaping-reviewer maintainers |
 | SHAPE-WRAPPER-BRIEF | One team’s single slice is wrapped by `docs/specs/rfc0099-migration-validation-record/fixtures/delivery-brief-flag-rename.md`. | Core-only | Fresh delivery-brief-mode review identifies the wrapper as `SHAPE-WRAPPER-BRIEF`; adjudicated evidence: `docs/specs/rfc0099-migration-validation-record/notes/shaping-efficacy-qa.md` | Core shaping-reviewer maintainers |
