+++ tests/roster/test_rfc0099_fixture_register.py
@@ -6,7 +6,7 @@ import hashlib
 import re
 from pathlib import Path
 
-REPOSITORY_ROOT = Path(__file__).resolve().parents[4]
+REPOSITORY_ROOT = Path(__file__).resolve().parents[2]
 REGISTER = (
     REPOSITORY_ROOT
     / "docs"
