+++ tools/lint-ci-parity.py
@@ -456,6 +456,8 @@ STEP_DISPOSITION: dict[str, tuple[str, str]] = {
         LOCAL("test-after-build-check"),
     "pytest shaping-review contracts (roster-owned)":
         LOCAL("test-after-build-check"),
+    "pytest RFC-0099 activation + fixture register (roster-owned)":
+        LOCAL("test-after-build-check"),
     "pytest TDD stub lifecycle contract (roster-owned)":
         LOCAL("test-after-build-check"),
     "pytest curation QA + RFC template contracts (roster-owned)":
